# docs/swagger.yaml
basePath: /
definitions:
...
  models.BranchRequest:
    properties:
      created_by:
        type: integer
      name:
        type: string
      updated_by:
        type: integer
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  models.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: helpdesk-dev.nopadol.com
info:
  contact: {}
  description: This is a sample swagger server in Go.
  title: Example API – Report Problem System
  version: "1.0"
paths:
...
  /api/v1/branch/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get branch details
      tags:
      - branches
  /api/v1/branch/create:
    post:
      consumes:
      - application/json
      description: Create a new branch
      parameters:
      - description: Branch data
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/models.BranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create branch
      tags:
      - branches
  /api/v1/branch/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete branch
      tags:
      - branches
  /api/v1/branch/list:
    get:
      consumes:
      - application/json
      description: Get list of all branches with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List branches
      tags:
      - branches
  /api/v1/branch/list/{query}:
    get:
      consumes:
      - application/json
      description: Search branches by query string with pagination
      parameters:
      - description: Search query (use 'all' for all branches)
        in: path
        name: query
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Search branches
      tags:
      - branches
  /api/v1/branch/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Branch data
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/models.BranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update branch
      tags:
      - branches
...
swagger: "2.0"